Clean transformer code
